Rescope and move classes in exo-player.forwarders where necessary.
